+++ algotrading/common/config/config.py
@@ -34,7 +34,7 @@ class Config:
 
         account = AccountConfigData(
             base_currency     = parser['account']['base_currency'],
-            leverage          = float(parser['account']['leverage']),
+            leverage          = int(parser['account']['leverage']),
             unit_size         = int(parser['account']['unit_size']),
             margin_call_level = float(parser['account']['margin_call_level']),
             stop_out_level    = float(parser['account']['stop_out_level'])
